:bug: fix: Lazy load Leaflet component and add suspense in Main
Imported the Leaflet component using 'React.lazy' for lazy loading and wrapped the Leaflet component with 'React.Suspense' to render a fallback while the component is being loaded. This solution resolves the issue of dependency on the window object during project compilation and build, allowing the code to execute correctly.